requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
diff-match-patch==20230430
Django==5.0.7
django-import-export==4.1.1
django-tailwind==3.8.0
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openpyxl==3.1.5
packaging==24.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
